edit file Bonding-centos

+++ Bonding-centos.sh
@@ -2,19 +2,26 @@
 
 #tao bien chua dia chi mac
 
-read -p "gateway:"gw
+read -p "gateway:" gw
 echo $gw
 
-read -p "MASK:"mask
+read -p "MASK:" mask
 echo $mask
 
 ipadd=`/sbin/ifconfig eth1 | grep inet | awk '{print $2}'`
+
+echo $ipadd
+
 #Tao card bond0
 
 touch /etc/sysconfig/network-scripts/ifcfg-bond0
 
+ifcfg-bond0=/etc/sysconfig/network-scripts/ifcfg-bond0
+
+
 #Cau hinh ip cho cong
-cat << EOF >> /etc/sysconfig/network-scripts/ifcfg-bond0
+cat << EOF >> $ifcfg-bond0
+
 DEVICE=bond0
 BOOTPROTO=none
 ONBOOT=yes
@@ -24,18 +31,21 @@ USERCTL=no
 BONDING_OPTS="mode=1 miimon=100"
 
 EOF
-
-cp /etc/modprobe.d/bonding.conf /etc/modprobe.d/bonding.conf.bka
+test -f /etc/modprobe.d/bonding.conf.bka || cp /etc/modprobe.d/bonding.conf /etc/modprobe.d/bonding.conf.bka
 ###Add 
 echo "alias bond0 bonding" >>/etc/modprobe.d/bonding.conf
 
-
-
 #### Cau hinh ip cho cac card
-cp /etc/sysconfig/network-scripts/ifcfg-eth1  /etc/sysconfig/network-scripts/ifcfg-eth1.bka
-cp /etc/sysconfig/network-scripts/ifcfg-eth0  /etc/sysconfig/network-scripts/ifcfg-eth0.bka
 
-cat << EOF >> /etc/sysconfig/network-scripts/ifcfg-eth1
+ifcfg-eth1=/etc/sysconfig/network-scripts/ifcfg-eth1
+
+test -f $ifcfg-eth1.bka || cp $ifcfg-eth1 $ifcfg-eth1.bka
+
+rm $ifcfg-eth1
+
+touch $ifcfg-eth1
+
+cat << EOF >> $ifcfg-eth1
 DEVICE=eth1
 MASTER=bond0
 SLAVE=yes
@@ -45,7 +55,13 @@ BOOTPROTO=none
 
 EOF
 #######################################################
-cat << EOF >> /etc/sysconfig/network-scripts/ifcfg-eth0
+ifcfg-eth0=/etc/sysconfig/network-scripts/ifcfg-eth0
+
+test -f $ifcfg-eth0.bka || cp $ifcfg-eth0 $ifcfg-eth0.bka
+rm $ifcfg-eth0
+touch $ifcfg-eth0
+
+cat << EOF >> $ifcfg-eth0
 
 DEVICE=eth0
 MASTER=bond0
